+++ src/otoole/read_strategies.py
@@ -15,7 +15,16 @@ logger = logging.getLogger(__name__)
 
 
 class ReadMemory(ReadStrategy):
-    """Read a dict of OSeMOSYS parameters from memory"""
+    """Read a dict of OSeMOSYS parameters from memory
+
+    Arguments
+    ---------
+    parameters : Dict[str, pd.DataFrame]
+        Dictionary of OSeMOSYS parameters
+    user_config : Dict[str, Dict]
+        User configuration
+
+    """
 
     def __init__(
         self, parameters: Dict[str, pd.DataFrame], user_config: Dict[str, Dict]
@@ -24,7 +33,7 @@ class ReadMemory(ReadStrategy):
         self._parameters = parameters
 
     def read(
-        self, filepath: Union[str, TextIO] = None, **kwargs
+        self, filepath: Union[str, TextIO, None] = None, **kwargs
     ) -> Tuple[Dict[str, pd.DataFrame], Dict[str, Any]]:
 
         config = self.user_config
@@ -58,6 +67,7 @@ class _ReadTabular(ReadStrategy):
 
         if "MODEOFOPERATION" in actual_headers:
             df = df.rename(columns={"MODEOFOPERATION": "MODE_OF_OPERATION"})
+            actual_headers = list(df.columns)
 
         if actual_headers[-1] == "VALUE":
             logger.info(
@@ -87,6 +97,11 @@ class _ReadTabular(ReadStrategy):
             except IndexError as ex:
                 logger.debug(f"Could not reshape {name}")
                 raise ex
+            except KeyError as ex:
+                logger.debug(
+                    f"Actual headers: {actual_headers}\nConverted headers: {converted_headers}"
+                )
+                raise ex
 
         all_headers = converted_headers + ["VALUE"]
         return narrow[all_headers].set_index(converted_headers)
@@ -111,7 +126,15 @@ class _ReadTabular(ReadStrategy):
 
 
 class ReadExcel(_ReadTabular):
-    """Read in an Excel spreadsheet in wide format to a dict of Pandas DataFrames"""
+    """Read in an Excel spreadsheet in wide format to a dict of Pandas DataFrames
+
+    Arguments
+    ---------
+    user_config : Dict[str, Dict]
+        User configuration
+    keep_whitespace : bool
+        Whether to keep whitespace in the dataframes
+    """
 
     def read(
         self, filepath: Union[str, TextIO], **kwargs
@@ -157,7 +180,15 @@ class ReadExcel(_ReadTabular):
 
 
 class ReadCsv(_ReadTabular):
-    """Read in a folder of CSV files"""
+    """Read in a folder of CSV files to a dict of Pandas DataFrames
+
+    Arguments
+    ---------
+    user_config : Dict[str, Dict]
+        User configuration
+    keep_whitespace : bool
+        Whether to keep whitespace in the dataframes
+    """
 
     def read(
         self, filepath, **kwargs
@@ -166,9 +197,13 @@ class ReadCsv(_ReadTabular):
         input_data = {}
 
         self._check_for_default_values_csv(filepath)
-        self._compare_read_to_expected(
-            names=[f.split(".csv")[0] for f in os.listdir(filepath)]
-        )
+        names = [
+            f.split(".csv")[0]
+            for f in os.listdir(filepath)
+            if f.split(".")[-1] == "csv"
+        ]
+        logger.debug(names)
+        self._compare_read_to_expected(names=names)
 
         default_values = self._read_default_values(self.user_config)
 
@@ -272,18 +307,36 @@ class ReadCsv(_ReadTabular):
 
 
 class ReadDatafile(ReadStrategy):
+    """Read in a datafile to a dict of Pandas DataFrames
+
+    Arguments
+    ---------
+    user_config : Dict[str, Dict]
+        User configuration
+    keep_whitespace : bool
+        Whether to keep whitespace in the dataframes
+
+    """
+
     def read(
         self, filepath, **kwargs
     ) -> Tuple[Dict[str, pd.DataFrame], Dict[str, Any]]:
 
         config = self.user_config
         default_values = self._read_default_values(config)
-        amply_datafile = self.read_in_datafile(filepath, config)
-        inputs = self._convert_amply_to_dataframe(amply_datafile, config)
-        for config_type in ["param", "set"]:
-            inputs = self._get_missing_input_dataframes(inputs, config_type=config_type)
-        inputs = self._check_index(inputs)
-        return inputs, default_values
+
+        # Check filepath exists
+        if os.path.exists(filepath):
+            amply_datafile = self.read_in_datafile(filepath, config)
+            inputs = self._convert_amply_to_dataframe(amply_datafile, config)
+            for config_type in ["param", "set"]:
+                inputs = self._get_missing_input_dataframes(
+                    inputs, config_type=config_type
+                )
+            inputs = self._check_index(inputs)
+            return inputs, default_values
+        else:
+            raise FileNotFoundError(f"File not found: {filepath}")
 
     def read_in_datafile(self, path_to_datafile: str, config: Dict) -> Amply:
         """Read in a datafile using the Amply parsing class
@@ -322,6 +375,7 @@ class ReadDatafile(ReadStrategy):
             elif attributes["type"] == "set":
                 elements += "set {};\n".format(name)
 
+        logger.debug("Amply Elements: %s", elements)
         return elements
 
     def _convert_amply_to_dataframe(