+++ CHANGELOG.rst
@@ -2,6 +2,38 @@
 Changelog
 =========
 
+(Development) Version 1.1.3
+===========================
+- Lock pandas to 2.1.4 or later
+- Capital Investment result calculation fixed
+
+Version 1.1.2
+=============
+- Update zenodo metadata for JOSS
+
+Version 1.1.1
+=============
+- Fixes CPLEX result processing docs
+- Added joss status badge to readme
+- Fix Tests on Windows
+- Update graphviz install instructions
+
+Version 1.1.0
+=============
+- Public Python API added to call otoole directly in Python files
+- ReadCplex directly reads in CPLEX solution files. Drops the need to transform and sort solution files
+- ReadGlpk class added to process GLPK solution files
+- Update to Pydantic v2.0
+- ReadResultsCbc renamed to ReadWideResults
+- Model validation instructions updated in documentation
+- The ``--input_datafile`` argument is deprecated, and the user now must supply the input data to process results
+- Locks pandas to <2.1
+
+Version 1.0.4
+=============
+- Fixed issue with pydantic v2.0.0
+- Dropped support for Python 3.8. Otoole now requires Python 3.9 or later
+
 Version 1.0.3
 =============
 - Improved error message for multiple names mismatches